+++ plugins/partiql-mockdb/src/main/kotlin/org/partiql/plugins/mockdb/LocalConnectorObject.kt
@@ -34,10 +34,6 @@ internal class LocalConnectorObject(
 
     public fun getDescriptor(): ValueDescriptor = descriptor
 
-    init {
-        println(descriptor)
-    }
-
     //
     //
     // HELPERS